Move to unified component instrumentation
